Add extended enumeration definitions to CuraEngine infill.
In response to CURA-10916, additional enumeration definitions were added, specifically for CuraEngine's infill patterns.
This update will ensure that every time a container is loaded, if it's a 'definition container' and not an 'extruder', all infill patterns present will extend the category with "UM Logo".
This will effectively link these infill patterns to the 'curaengine_plugin_infill_generate' plugin.
Be mindful there is also a pending task to only start the plugin if the extended enums are actually being used.

CURA-10916