Fixed retrieving and storing BroadcastSettings
A new file metadata.h was created to store plugin information such as slot_version, plugin_name, and plugin_version. This change simplifies code by ensuring these data items are stored in a central place. It also reduces redundancy by eliminating the replication of these variables across the different plugin services (handshake, broadcast, generate). These services have been updated to use the new Metadata struct, with the plugin's metadata object being initialized when the plugin is created.

Additionally, the getUuid() function was moved to metadata.h to be utilized by multiple services, and the handling of client metadata was simplified in generate and broadcast services.

In settings.h, the getPattern() function was revised to include the plugin_name as a parameter, to match the plugin name stored in metadata for the corresponding pattern.

Finally, in main.cpp, the handshake and generate services were updated to use the plugin's metadata when they are added to the plugin.

This change improves the structure and maintainability of the plugin code, whilst enhancing its consistency and making it easier to update or modify plugin information in the future.

Contributes to CURA-10916