Make retrieving settings more robust
This commit refactors how we fetch pattern and Tile type from request
settings in the Settings and Generate classes. Now, if a setting can't
be found or retrieved from request settings, it is logged as an error and
the process terminates early. This makes the code simpler and easier to
read, and more robust as it will now fail with a clear error message
when a setting is not present.

It also includes minor adjustments to clean the Settings class, mainly
implementing early terminations on error and avoiding unnecessary
computations.

Contributes to CURA-10619